// src/exceptions/exception-message-code.enum.ts
export enum ExceptionMessageCode {
  MISSING_TOKEN = 'MISSING_TOKEN',
  EXPIRED_TOKEN = 'EXPIRED_TOKEN',
  INVALID_TOKEN = 'INVALID_TOKEN',
  EMAIL_OR_PASSWORD_INVALID = 'EMAIL_OR_PASSWORD_INVALID',
  USER_EMAIL_EXISTS = 'USER_EMAIL_EXISTS',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  REFRESH_TOKEN_REUSE = 'REFRESH_TOKEN_REUSE',
  RECOVER_PASSWORD_REQUEST_NOT_FOUND = 'RECOVER_PASSWORD_REQUEST_NOT_FOUND',
  RECOVER_PASSWORD_REQUEST_INVALID_CODE = 'RECOVER_PASSWORD_REQUEST_INVALID_CODE',
  RECOVER_PASSWORD_REQUEST_TIMED_OUT = 'RECOVER_PASSWORD_REQUEST_TIMED_OUT',
  RECOVER_PASSWORD_REQUEST_INVALID = 'RECOVER_PASSWORD_REQUEST_INVALID',
  LEGAL_DOCUMENT_NOT_FOUND = 'LEGAL_DOCUMENT_NOT_FOUND',
  LEGAL_DOCUMENT_PARAGRAPH_NOT_FOUND = 'LEGAL_DOCUMENT_PARAGRAPH_NOT_FOUND',
  USER_DETAILS_NOT_FOUND = 'USER_DETAILS_NOT_FOUND',
  MISSING_CURRENT_USER_PAYLOAD = 'MISSING_CURRENT_USER_PAYLOAD',
  ACCOUNT_VERIFFICATION_REQUEST_NOT_FOUND = 'ACCOUNT_VERIFFICATION_REQUEST_NOT_FOUND',
  ACCOUNT_VERIFFICATION_REQUEST_INVALID_CODE = 'ACCOUNT_VERIFFICATION_REQUEST_INVALID_CODE',
  ACCOUNT_VERIFFICATION_REQUEST_TIMED_OUT = 'ACCOUNT_VERIFFICATION_REQUEST_TIMED_OUT',
  FRIENDS = 'FRIENDS',
  FRIEND_REQUEST_RECEIVED = 'FRIEND_REQUEST_RECEIVED',
  FRIEND_REQUEST_SENT = 'FRIEND_REQUEST_SENT',
  FRIEND_REQUEST_NOT_FOUND = 'FRIEND_REQUEST_NOT_FOUND',
  CANT_FRIEND_YOURSELF = 'CANT_FRIEND_YOURSELF',
  NOT_FRIENDS = 'NOT_FRIENDS',
  INVALID_PARTICIPANT_COUNT = 'INVALID_PARTICIPANT_COUNT',
  ROOM_NOT_FOUND = 'ROOM_NOT_FOUND',
  CHAT_NOT_FOUND = 'CHAT_NOT_FOUND',
  JOKER_ROOM_DETAILS_EXISTS = 'JOKER_ROOM_DETAILS_EXISTS',
  JOKER_ROOM_DETAILS_NOT_FOUND = 'JOKER_ROOM_DETAILS_NOT_FOUND',
  USER_NOT_IN_CURRENT_ROOM = 'USER_NOT_IN_CURRENT_ROOM',
  ROUND_EXISTS_WITH_SAME_ORDER = 'ROUND_EXISTS_WITH_SAME_ORDER',
  ROUND_CARDS_EXISTS = 'ROUND_CARDS_EXISTS',
  TAKEN_ROUND_CARDS_EXISTS = 'TAKEN_ROUND_CARDS_EXISTS',
  PASSED_ROUND_CARDS_EXISTS = 'PASSED_ROUND_CARDS_EXISTS',
  INVALID_CARD_GROUP_COUNT = 'INVALID_CARD_GROUP_COUNT',
  JOKER_GAME_EVENT_EXISTS = 'JOKER_GAME_EVENT_EXISTS',
  TRUMP_CALLED = 'TRUMP_CALLED',
  USER_DOES_NOT_BELONG_TO_ROOM = 'USER_DOES_NOT_BELONG_TO_ROOM',
  JOKER_ROUND_SCORE_EXISTS = 'JOKER_ROUND_SCORE_EXISTS',
  ROOM_PARTICIPANT_NOT_FOUND = 'ROOM_PARTICIPANT_NOT_FOUND',
  NOT_YOUR_TURN = 'NOT_YOUR_TURN',
  ROOM_PARTICIPANT_EXISTS = 'ROOM_PARTICIPANT_EXISTS',
  JOKER_MOVE_EXISTS = 'JOKER_MOVE_EXISTS',
  ROOM_NOT_ACTIVE = 'ROOM_NOT_ACTIVE',
  ROUND_NOT_FOUND = 'ROUND_NOT_FOUND',
  INVALID_ROUND_STATUS = 'INVALID_ROUND_STATUS',
  INVALID_JOKER_MOVE = 'INVALID_JOKER_MOVE',
  CARD_NOT_FOUND = 'CARD_NOT_FOUND',
  ROUND_CARDS_NOT_FOUND = 'ROUND_CARDS_NOT_FOUND',
  INVALID_GUESS_CALL = 'INVALID_GUESS_CALL',
}